Refactored Domain to use a feature based approach
