+++ src/Application/Features/TodoItems/Commands/CreateTodoItem/CreateTodoItemCommand.cs
@@ -1,5 +1,5 @@
 ﻿using SSW.CleanArchitecture.Application.Common.Interfaces;
-using SSW.CleanArchitecture.Domain.Entities;
+using SSW.CleanArchitecture.Domain.TodoItems;
 
 namespace SSW.CleanArchitecture.Application.Features.TodoItems.Commands.CreateTodoItem;
