+++ src/Application/Features/TodoItems/Specifications/AllTodoItemSpec.cs
@@ -1,5 +1,5 @@
 using Ardalis.Specification;
-using SSW.CleanArchitecture.Domain.Entities;
+using SSW.CleanArchitecture.Domain.TodoItems;
 
 namespace SSW.CleanArchitecture.Application.Features.TodoItems.Specifications;
