--- src/Domain/GlobalUsing.cs
@@ -1,2 +1,0 @@
-﻿global using SSW.CleanArchitecture.Domain.Common;
-global using SSW.CleanArchitecture.Domain.Enum;
