+++ tests/Application.UnitTests/Features/TodoItems/Specifications/AllTodoItemSpecTests.cs
@@ -1,6 +1,6 @@
 using Bogus;
 using SSW.CleanArchitecture.Application.Features.TodoItems.Specifications;
-using SSW.CleanArchitecture.Domain.Entities;
+using SSW.CleanArchitecture.Domain.TodoItems;
 
 namespace SSW.CleanArchitecture.Application.UnitTests.Features.TodoItems.Specifications;
