+++ tests/Application.UnitTests/Features/TodoItems/Specifications/TodoItemByTitleSpecTests.cs
@@ -1,5 +1,5 @@
 using SSW.CleanArchitecture.Application.Features.TodoItems.Specifications;
-using SSW.CleanArchitecture.Domain.Entities;
+using SSW.CleanArchitecture.Domain.TodoItems;
 
 namespace SSW.CleanArchitecture.Application.UnitTests.Features.TodoItems.Specifications;
